+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- ir: Add various `unit()`, `unit_mut()`, `get_unit()`, and `get_unit_mut()` functions.
+
 ## 0.10.0 - 2019-11-24
 ### Added
 - Add Sublime Text syntax highlighting.